feat(RUN-819): extract worker_proxies from worker.py
Move ProxiedLLMClient, ProxiedRunsightTeamRunner, create_llm_client,
create_runner, and create_tool_stubs to worker_proxies.py. No re-exports,
no __getattr__ shim. Worker accesses moved symbols via _proxies module alias.
